Final checkin: ABANDONED
It just felt like more work for the player than it was worth.
So much back and forth and writing and trying to find where you should record it, all just to record it on the character sheet at the end.
Sure, the 2024 PHB character creation section makes it a little difficult to figure out all of the stuff you get and how to combine it, but it's not that difficult.